IE-extension-backend/requirements.txt
openssl
cryptography
numpy
scipy
openai-clip
pillow
jpeglib
requests-toolbelt